fix: Files without path
When replicating files from CouchDB, we do not retrieve path, as it is
not persisted in database.
Therefore, we dynamically compute the path when a search returns files
results.
Note we do it at this moment rather than at replication time, because
this operation requires to query the parent directory, which can be done
directly in-memory through the store once the search index is built,
rather than an additional server query for each replicated file.