feat: Use local search only for persistent sessions
We now rely on the checkbox on login to determine whether or not the
local search should be used. If the user enable this checkbox, it means
he trusts this device and accept to store local data on it.
Otherwise, we do not store any private data.

See https://github.com/cozy/cozy-libs/pull/2611 for the search part.