fix: Handle reset for old browser
We rely on the `indexedDb.databases()` method to iterate over all local
db to be removed, at login time. We also do it later, in the
`SharedWorkerProvider` as a fallback.
However, this does not work for old browsers that do not support this
function.
Therefore, we now store the db names locally in PouchLink, and rely on
it to remove the databases.